DEV: restructure tests: move Surveys tests
