+++ tests/acceptance/surveys/AjaxModeTest.php
@@ -1,11 +1,9 @@
 <?php
-
-namespace ls\tests;
+namespace ls\tests\acceptance\surveys;
 
 use Facebook\WebDriver\WebDriverBy;
-use Facebook\WebDriver\WebDriverExpectedCondition;
 use Facebook\WebDriver\Exception\NoSuchElementException;
-
+use ls\tests\TestBaseClassWeb;
 /**
  * @since 2017-11-13
  * @group ajaxmode
